# Accelerate CTC greedy decoding by around 10%