--- a/nemo/collections/asr/models/ctc_models.py
+++ b/nemo/collections/asr/models/ctc_models.py
@@ -666,9 +666,7 @@ class EncDecCTCModel(ASRModel, ExportableEncDecModel, ASRModuleMixin, InterCTCMi
             # See comment in
             # ctc_greedy_decoding.py::GreedyCTCInfer::forward() to
             # understand this idiom.
-            logits_cpu = torch.empty(logits.shape,
-                                     dtype=logits.dtype,
-                                     pin_memory=True)
+            logits_cpu = torch.empty(logits.shape, dtype=logits.dtype, pin_memory=True)
             logits_cpu.copy_(logits)
             logits_len = logits_len.cpu()
             # dump log probs per file