@galv 2024-02-29
@titu1994 FYI, I made a fix. I tested by simply running running

```
python examples/asr/speech_to_text_eval.py  model_path=$HOME/.cache/huggingface/hub/models--nvidia--parakeet-ctc-1.1b/snapshots/085a3de63c7598065b072cd8f2182e6a5fa593eb/parakeet-ctc-1.1b.nemo  dataset_manifest=/home/dgalvez/scratch/data/test_other.json  batch_size=16  output_filename=test_other_decoded.jsonl  amp=true  amp_dtype=bfloat16  use_cer=false num_workers=1
```

with and without "CUDA_VISIBLE_DEVICES="

But the existing CTC test cases would have caught the error anyway if jenkins had run successfully, if jenkins had worked instead of getting stuck on fetching the git repo.

I also double checked my throughput results. They are a little bit better than originally reported because of a dumb error on my part. We are going from 28.75 seconds (RTFx 676) to 25.5 seconds (RTFx 762) after warmup on librispeech test other. Which is a nice speedup.